fix x86_64 macOS crash by using an alternative way to embed files
I think there's a bug in the Zig MachO linker that means that assembly
code around `@embedFile` is manged in the codesigning process. This is
the issue we made about it https://github.com/ziglang/zig/issues/23704.

To workaround this, we have converted all the files into C++ source and
use that instead. https://github.com/floe-audio/floe-embedded-files-workaround
It's in a separate repo because we don't want to check in these large
files when this hopefully should just be a temporary workaround.