add rudimentary protection against hosts trying to have multiple main threads
We can somewhat protect against hosts trying to call our functions from
multiple 'main' threads by switching an atomic flag for the duration
of a main-thread function. If the flag is already set - the host is
misbehaving and we reject it.

Additionally, I've moved some of the entering of main thread further up
so that this multi-thread protection happens before accessing variables.