build: use --global-cache-dir to make paths relative build root
Rather than using the default global cache (which is in the $HOME dir),
we now say that the global cache is within the build root, this way,
our ffile-prefix-map maps all paths to be non-absolute and therefore
independent of what machine or where the source code is. Our reports
in Sentry will be clearer and now we no longer have /home/runner/.cache
paths in the debug info of the binary.