fix scan-folder troubles
Closes #142

Whilst unable to recreate the issue #142, identified a problem: we were
using an AtomicRefList for the scan-folders and passing a 'node' to
the thread pool to do the actual scanning. Nodes in the atomic ref lists
are like slots - their content may change. I believe that what was
happening was that the content inside the slot changed, so that we were
trying to scan a folder that was already scanned - resulting in the
assertion in the issue firing.

The fix is to use a simpler, more appropriate data structure for this.
We now have an mutex protected array.