add ability to check for updates and show on the GUI
At startup, Floe will check online if a newer version is available. If
so, it will display a small indicator on the Info icon button. More
details will be shown on the Info panel. Can be disabled in preferences.
Each new version can also be ignored by pressing a button.