refactor: unify param and tempo changes and note events into single call
The goal here is to have a clear interface for processors for processing
at the 2 rate we want to use: audio-rate and control-rate.

Control-rate being the start of each subblock - this is granularity that
parameter changes will occur due to them being too expensive to adapt on
every parameter. As well as parameters changes, tempo changes and MIDI
events are processed at this rate. This patch creates this unified
interface ready for upcoming sub-blocking.