try workaround tsan false positives
The tsan output when we hit these false positives suggest that tsan
doesn't understand that memory is invalidated when its unmapped. This
theory checks out because we don't always hit the error since virtual
memory is only sometimes reused, and that the corressponding read/write
that its complaining about seem to be in completely separate parts of
the program which just both happen to use page allocations.

The solution we are trying is to use malloc/free for page allocations
since presumably, tsan has special understand of memory
acquired/released in this manner whereas it does not for mmap/munmap.

Hopefully closes #161.