tests: add gha-annotations-output-path to tests instead of step summary
Writing the step summary file has some problems: it didn't work at all
on Windows for some reason, and it had potential concurrent write issues
because even though it used file locks, these are only hints.

Instead, the better solution to this is to use github actions
annotations with pop up in a designated section on the web UI. Getting
stdout from application is a bit tricky with zig build, since zig build
tends to discard stdout info; Zig takes the stance that stdout is for
piping, not printing to terminal. So rather than try to fight this
approach, we offer a separate arg to the test program the writes
annotations to a file - the caller can then put this information
wherever it likes (stdout, a buffer, etc.).