fix sample library server spuriously returning NotFound for libraries
Here's what I think was happening: the server thread had a private copy
of the scan-folders list and would not see folders added via
SetExtraScanFolders until its next tick. In other words within the
current tick it could see 'all my scan-folders are scanned and I can't
find the requested library - it's an error', when in fact there is an
additional scan-folder but it's not not in the local copy.

To fix this, I chose to restructure the scan-folders so that it's a
fixed size array (we know the max allowed), and just mutex-protect
each folder.

Tests pass but we have a nasty 'while(not_ready) sleep()' that we need
to replace with something more responsive and less wasteful such as
a condition variable perhaps.