fix issue where param queue to audio thread filled up
The error reported shows this can occasionally happen in Reaper and
FL Studio. Perhaps the host is issuing many re-loads of parameter values
without the audio thread pumping.

We fix this by using a more robust method where the parameter value
changes are represented in a fixed size array of atomics.