preset server: fix removing all folder before rescan
We'd remove-and-publish all folders when a scan-folder needed a rescan.
Then we'd build them up again one-by-one. That left a window where
readers could see empty folders - not appropriate considering the common
case is that only some files change in a rescan. We now replace folders
one-by-one, and remove any folders that no longer exist in reality.